IFS-2016-PF-SpendProfile-Action
REST Layer changes to use JPQL.

+++ ifs-rest-api-service/src/main/java/org/innovateuk/ifs/competition/service/CompetitionPostSubmissionRestService.java
@@ -25,7 +25,7 @@ public interface CompetitionPostSubmissionRestService {
 
     RestResult<List<CompetitionPendingSpendProfilesResource>> getPendingSpendProfiles(long competitionId);
 
-    RestResult<Integer> countPendingSpendProfiles(long competitionId);
+    RestResult<Long> countPendingSpendProfiles(long competitionId);
 
     RestResult<Void> closeAssessment(long competitionId);
